fix: use uint64 for string_offset in tp_segment_read_term_at_index
header->strings_offset is uint64; using uint32 silently truncated
the computed offset for segments larger than 4GB. The old segment.c
version was correct (uint64), but merge.c's version had the bug.
The shared function inadvertently copied the buggy type.

+++ src/segment/dictionary.c
@@ -143,7 +143,7 @@ tp_segment_read_term_at_index(
 		uint32			*string_offsets,
 		uint32			 index)
 {
-	uint32 string_offset;
+	uint64 string_offset;
 	uint32 length;
 	char  *term;
